LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-delaynode-interface/no-dezippering-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.

Harness Error (FAIL), message = Unhandled rejection: The object is in an invalid state.

PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "test0" 
PASS Executing "test1" 
PASS Executing "test2" 
PASS Audit report 
PASS > [test0] Test DelayNode has no dezippering 
PASS   output[0:63] contains only the constant 0. 
PASS   output[64:127] is identical to the array [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16...]. 
FAIL X output[128:] expected to be equal to the array [113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128...] but differs in 16256 places:
	Index	Actual			Expected
	[0]	6.5000000000000000e+1	1.1300000000000000e+2
	[1]	6.6000000000000000e+1	1.1400000000000000e+2
	[2]	6.7000000000000000e+1	1.1500000000000000e+2
	[3]	6.8000000000000000e+1	1.1600000000000000e+2
	...and 16252 more errors. assert_true: expected true got false
FAIL < [test0] 1 out of 3 assertions were failed. assert_true: expected true got false
PASS > [test1] Test value setter and setValueAtTime 
FAIL X .value setter output does not equal [0,0,0,0,0.05180763825774193,0.10919986665248871,0.16622987389564514,0.22270867228507996,0.27844879031181335,0.33326560258865356,0.38697704672813416,0.4394051730632782,0.4903762638568878,0.5397207736968994,0.5872756242752075,0.6328827142715454...] with an element-wise tolerance of {"absoluteThreshold":0.000065819,"relativeThreshold":0}.
	Index	Actual			Expected		AbsError		RelError		Test threshold
	[384]	5.2849482744932175e-2	5.2146881818771362e-1	4.6861933544278145e-1	8.9865265016496554e-1	6.5819000000000001e-5
	[385]	-4.7132032923400402e-3	5.6970381736755371e-1	5.7441702065989375e-1	1.0082730765507566e+0	6.5819000000000001e-5
	[386]	-6.2260106205940247e-2	6.1604958772659302e-1	6.7830969393253326e-1	1.1010634654195592e+0	6.5819000000000001e-5
	[387]	-1.1960060894489288e-1	6.6035234928131104e-1	7.7995295822620392e-1	1.1811163526185062e+0	6.5819000000000001e-5
	[388]	-1.7654430866241455e-1	7.0246493816375732e-1	8.7900924682617188e-1	1.2513211678918825e+0	6.5819000000000001e-5
	...and 15995 more errors.
	Max AbsError of 1.9097955822944641e+0 at index of 13007.
	[13007]	9.5548820495605469e-1	-9.5430737733840942e-1	1.9097955822944641e+0	2.0012373661209018e+0	6.5819000000000001e-5
	Max RelError of 5.4351310197854377e+2 at index of 429.
	[429]	-5.6675183773040771e-1	1.0446786182001233e-3	5.6779651634860784e-1	5.4351310197854377e+2	6.5819000000000001e-5
 assert_true: expected true got false
FAIL X .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test1] 2 out of 2 assertions were failed. assert_true: expected true got false
PASS > [test2] Test value setter and modulation 
FAIL X With modulation: .value setter output does not equal [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0...] with an element-wise tolerance of {"absoluteThreshold":0,"relativeThreshold":0}.
	Index	Actual			Expected		AbsError		RelError		Test threshold
	[384]	5.2863571792840958e-2	5.2145665884017944e-1	4.6859308704733849e-1	8.9862326830686223e-1	0.0000000000000000e+0
	[385]	9.4109612703323364e-1	-5.8518588542938232e-1	1.5262820124626160e+0	2.6082003179941720e+0	0.0000000000000000e+0
	[386]	6.8975055217742920e-1	-9.7813159227371216e-1	1.6678821444511414e+0	1.7051715307283677e+0	0.0000000000000000e+0
	[387]	-3.4253859519958496e-1	-2.4920801818370819e-1	9.3330577015876770e-2	3.7450872446277572e-1	0.0000000000000000e+0
	[388]	-9.8527598381042480e-1	7.1552461385726929e-1	1.7008005976676941e+0	2.3769980301571634e+0	0.0000000000000000e+0
	...and 9668 more errors.
	Max AbsError of 1.9098862409591675e+0 at index of 16040.
	[16040]	-9.5888763666152954e-1	9.5099860429763794e-1	1.9098862409591675e+0	2.0082955246498160e+0	0.0000000000000000e+0
	Max RelError of 6.8668017258543594e+4 at index of 4280.
	[4280]	-5.6597548723220825e-1	8.2423193816794083e-6	5.6598372955158993e-1	6.8668017258543594e+4	0.0000000000000000e+0
 assert_true: expected true got false
FAIL X With modulation: .value setter output matches setValueAtTime output is not true. Got false. assert_true: expected true got false
FAIL < [test2] 2 out of 2 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 3 out of 3 tasks were failed. assert_true: expected true got false
